LayoutTests/imported/w3c/web-platform-tests/referrer-policy/no-referrer/attr-referrer/same-origin/http-https/script-tag/keep-origin-redirect/generic.http-expected.txt


FAIL The referrer URL is omitted when a
                                 document served over http requires an https
                                 sub-resource via script-tag using the attr-referrer
                                 delivery method with keep-origin-redirect and when
                                 the target request is same-origin. assert_equals: Reported Referrer URL is 'omitted'. expected (undefined) undefined but got (string) "http://localhost:8800/referrer-policy/no-referrer/attr-referrer/same-origin/http-https/script-tag/keep-origin-redirect/generic.http.html"
